warehouse_rest/requirements.txt
Django==1.10
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
psycopg2==2.7.4
psycopg2-binary==2.7.4